feat(spyrun): add init command
Add the `init` command to sync spyrun.toml and other files to the remote system.

This command will copy `spyrun.toml`, `common.ps1`, `launch.js`, and `init.ps1` to the remote system. It also will copy the entire `system` and `user` configuration folders from the remote system to the local system.

This command will greatly simplify the process of setting up the spyrun system on a new machine.